Created file set.pl with initial implementation of containsSet
File contains code and comments for containsSet.

Given a positive integer and two lists, returns true if
the length of the second list is equal to the integer,
the second list is a sublist of the first, and every list
which is an element of the second list is compatible
with every other list in the second list.

Two lists are compatible if every element in them
differs.

The basic functionality of the assignment has been
implemented, as well as having the compatibility
function work for lists of arbitrary length instead of
only lists of one set length.

Additional functionality to be added is determining
compatibility using the rules of the game Set, and
allowing for the case where list elements of the second
list are out of order from how the occur in the first list.